requirements.txt
django
python-binance
pandas
jsonschema
requests
psycopg2
pyrsistent
pytest
mock
responses
tables
torch
git+https://github.com/Binance-docs/Binance_Futures_python
sklearn
xgboost
